+++ app/Http/Controllers/Admin/Base/RoleController.php
@@ -97,6 +97,9 @@ class RoleController extends BaseAdminController
      */
     public function destroy($id)
     {
+        if(1 == $id){
+            return $this->ajax_return('500', '操作失败！包含保护项！！');
+        }
         $result = Role::destroy($id);
         // $this->log(__CLASS__, __FUNCTION__, $request, "添加菜单");
         return $this->ajax_return(200, '操作成功！');
